+++ src/wgChannelMap.cc
@@ -360,7 +360,7 @@ bool wgChannelMap::GetViewPlnCh(const int dif_id, const int chip_id, const int c
 }; //GetViewPlnCh
 
 //******************************************************************************
-bool wgChannelMap::GetXYZ(int view, int pln, int ch, float& x, float& y, float& z)
+bool wgChannelMap::GetXYZ(int view, int pln, int ch, double& x, double& y, double& z)
 {
 
   unsigned short grid = (ch/20)&0b11;
@@ -451,7 +451,7 @@ Map_t wgChannelMap::load_mapping(){
   Map_t map_struct;
   //reading mapping
   vector<int> pln, ch, grid;
-  vector<float> x, y, z;
+  vector<double> x, y, z;
   for(size_t idif=0;idif<NumDif;idif++) {
     for(size_t ichip=0;ichip<NumChip;ichip++) {
 	  wrapArrayInVector( map_struct.pln [idif][ichip].data(), map_struct.pln [idif][ichip].size(), pln );
@@ -495,7 +495,7 @@ MapInv_t wgChannelMap::load_mapping_inv(size_t n_chans){
     for(int ichip=0;ichip<NumChip;ichip++){
       int view;
       vector<int> pln(n_chans), ch(n_chans), grid(n_chans);
-      vector<float> x(n_chans), y(n_chans), z(n_chans);
+      vector<double> x(n_chans), y(n_chans), z(n_chans);
       this->GetMap(idif, ichip, view, pln, ch, grid, x, y, z);
         
       for(int ich=0;ich<NumChipCh;ich++){
